refactor(api)!: 🏗️ 统一 serviceId 数据类型为 Long 以支持更大 ID 范围
后端所有涉及 serviceId 的接口 / 实体 / Mapper / VO / DTO 均由 Integer 调整为 Long，
避免主键过大溢出；同步更新前端 TypeScript 参数及接口逻辑，保持全栈一致性。

- AutoScaleTaskEntity、ClusterServiceRoleInstanceVO、FrameServiceRoleDTO/VO
  等模型字段类型调整
- KubernetesDashboardService、Controller、ServiceImpl、Mapper 方法签名对应修改
- ts api 封装及 Pods/Services Dashboard 组件新增可选 serviceId 透传
- 移除控制台多余日志，简化刷新依赖项，提高组件稳定性

破坏性变更：所有通过 Query&Path 传入的 serviceId 现仅接受 Long 类型。

+++ datasophon-api/src/main/java/com/datasophon/api/controller/v1/kubernetes/KubernetesDashboardController.java
@@ -81,7 +81,7 @@ public class KubernetesDashboardController {
     @RequestMapping("/deployments")
     public Result<PageVO<KubernetesResourceVO>> getDeployments(
             @ClusterId Long clusterId,
-            @RequestParam(value = "serviceId", required = false) Integer serviceId,
+            @RequestParam(value = "serviceId", required = false) Long serviceId,
             @RequestParam(value = "namespace", required = false) String namespace,
             @RequestParam(value = "pageNum", required = false, defaultValue = "1") Integer pageNum,
             @RequestParam(value = "pageSize", required = false, defaultValue = "10") Integer pageSize) {
@@ -112,7 +112,7 @@ public class KubernetesDashboardController {
      */
     @GetMapping("/pods")
     public Result<PageVO<KubernetesResourceVO>> getPodsInfo(@ClusterId Long clusterId,
-            @RequestParam(name = "serviceId", required = false) Integer serviceId,
+            @RequestParam(name = "serviceId", required = false) Long serviceId,
             @RequestParam(name = "namespace", required = false) String namespace,
             @RequestParam(name = "pageNum", required = false, defaultValue = "1") Integer pageNum,
             @RequestParam(name = "pageSize", required = false, defaultValue = "10") Integer pageSize) {
@@ -319,7 +319,7 @@ public class KubernetesDashboardController {
     @RequestMapping("/daemonsets")
     public Result<PageVO<KubernetesResourceVO>> getDaemonSets(
             @ClusterId Long clusterId,
-            @RequestParam(value = "serviceId", required = false) Integer serviceId,
+            @RequestParam(value = "serviceId", required = false) Long serviceId,
             @RequestParam(value = "namespace", required = false) String namespace,
             @RequestParam(value = "pageNum", required = false, defaultValue = "1") Integer pageNum,
             @RequestParam(value = "pageSize", required = false, defaultValue = "10") Integer pageSize) {
@@ -408,7 +408,7 @@ public class KubernetesDashboardController {
     @RequestMapping("/jobs")
     public Result<PageVO<KubernetesResourceVO>> getJobs(
             @ClusterId Long clusterId,
-            @RequestParam(value = "serviceId", required = false) Integer serviceId,
+            @RequestParam(value = "serviceId", required = false) Long serviceId,
             @RequestParam(value = "namespace", required = false) String namespace,
             @RequestParam(value = "pageNum", required = false, defaultValue = "1") Integer pageNum,
             @RequestParam(value = "pageSize", required = false, defaultValue = "10") Integer pageSize) {
@@ -509,7 +509,7 @@ public class KubernetesDashboardController {
     @RequestMapping("/resource-stats")
     public Result<K8sResourceStatsDTO> getResourceStats(
             @ClusterId Long clusterId,
-            @RequestParam(value = "serviceId", required = false) Integer serviceId,
+            @RequestParam(value = "serviceId", required = false) Long serviceId,
             @RequestParam(value = "namespace", required = false) String namespace) {
         try {
             // 添加调试日志
